Use relative imports for interfaces
When importing Flexible Voting into another repository, the forge
coverage tool was failing with the error: "Could not read source
code for analysis" linked to the imports of interface files from
the FlexibleVotingClient contract. While the non-relative imports
work fine in the context of this repo, and even in the context of
running tests normally when used as an import, they cause an issue
for the coverage tool.

It's unclear if this is a bug with foundry or something with our
configuration. Regardless, the easiest fix is simply to use relative
imports, which we do here.

Read more about the issue here:
https://github.com/foundry-rs/foundry/issues/2915