Switch to custom errors (#76)
Current best practice is to throw custom error types rather than strings on revert cases, and to append the contract name to the front of custom error types with `__` in between. This PR updates all revert cases in the FlexibleVotingClient.sol to use this convention. Since you cannot throw custom error types in require statements except in the latest version of solidity, this PR also converts from require statements to conditionals.