Fix interfaces, use token when asking about balances, not governor
It was weird to begin with to have functions like `rawBalanceOf` and
`deposit` and `withdraw` take a governor argument instead of a token
argument. It's much more natural to think in terms of
depositing/withdrawing/having a balance of a *token* than a governor.

Also it quickly became clear in tests that this weirdness was more than
inconvenient, as there were contexts where a deposit would need to be
made, the token was available, but the governor was not -- and since you
can't ask the token for the governor, we were stuck.